Transaction table vamount spent data type changed from float to decimal

+++ db/spending_tracker.sql
@@ -18,7 +18,7 @@ CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
     date VARCHAR(255),
     time VARCHAR(255),
-    amount_spent FLOAT,
+    amount_spent DECIMAL,
     merchant_id INT REFERENCES merchants(id) ON DELETE CASCADE,
     tag_id INT REFERENCES tags(id) ON DELETE CASCADE
 );
